TriOsc
had to include #include "pq_map_real.h"
#include "pq_time.h" in TriOsc.cpp
