+++ src/PqExtra.h
@@ -29,61 +29,6 @@
 #include "MovingStats.h"
 #include "SimpleStats.h"
 
-/**
- * Triangle/sawtooth oscillator.
- */
-class TriOsc : public PqGetter {
-public:
-  /**
-   * Constructor.
-   * @param period the period of oscillation (in seconds)
-   * @param width a value in [0, 1] that determines the point at which the wave reaches its maximum point (expressed as a fraction of the period)
-   */
-  TriOsc(float period=1.0f, float width=0.5f);
-  virtual ~TriOsc() {}
-
-  /// Returns value in [0, 1].
-  virtual float get() { return _value; }
-
-  /**
-   * Sets the period (in seconds).
-   * @param period the period of oscillation (in seconds)
-   * @return the unit itself
-   */
-  virtual TriOsc& period(float period);
-
-  /**
-   * Sets the frequency (in Hz).
-   * @param frequency the frequency of oscillation (in Hz)
-   * @return the unit itself
-   */
-  virtual TriOsc& frequency(float freq) { return period(1/freq); }
-
-  /**
-   * Sets the width of the wave.
-   * @param width a value in [0, 1] that determines the point at which the wave reaches its maximum point (expressed as a fraction of the period)
-   * @return the unit itself
-   */
-  virtual TriOsc& width(float width);
-
-protected:
-  // Core Plaquette methods.
-  virtual void setup();
-  virtual void update();
-
-  // Current value of the signal.
-  float _value;
-
-  // Period (seconds).
-  float _period;
-
-  // Tipping-point (in % of period).
-  float _width;
-
-  // Start time of each period (in seconds).
-  float _startTime;
-};
-
 // TODO: implement a floating-point version of Chrono in Plaquette and make
 // ramp a subclass of that class.
 /**
